Deal with albums that are different pressing.
If specific pressing of album is not in database, not a single track can be verified by AR.
This only accepts albums that are ripped without errors.
Erroreus rips are caught by checking "There were errors" which returns error code 3.

This works, but more checks or better logging would be more than nice.